Rename BC Custom Canvas to BC Marine Canvas
Originally, this project was going to be named 'BC Custom Canvas',
because although the business' name is 'BC Marine Canvas', that URL
wasn't available. We thought `bccustomcanvas.com` would be a good
alternative. However, that decision has been reconsidered, and the URL
will actually be `bcmarinecanvasri.com`.

This commit renames everything to reflect the new URL.